Pull modeForRoute into disruptions and test it
